Formed the final form of the repository
Originally, I wanted this repository's purpose to be to have a script to
automatically setup a debian installation. However, after some trial and
error having an automated script is too generic and complicated to
actually be feasible. Now this repository stores my config (dot) files
and a list of installed packages on my system.